requirements.txt
Django==5.2
djangorestframework==3.14.0
pymysql~=1.1.1
gunicorn==20.1.0
razorpay~=1.4.2